requirements.txt
numpy
scipy
pandas
matplotlib
astropy
astroquery
pydantic
PyYAML
rich
pytest
